// client/src/pages/mill-staff/sales-reports/gunny-sales/data/schema.ts
import { z } from 'zod'

export const gunnySalesSchema = z.object({
    date: z.string().min(1, 'Date is required'),
    partyName: z.string().optional(),
    newGunnyQty: z.number().optional(),
    newGunnyRate: z.number().optional(),
    oldGunnyQty: z.number().optional(),
    oldGunnyRate: z.number().optional(),
    plasticGunnyQty: z.number().optional(),
    plasticGunnyRate: z.number().optional(),
})

export type GunnySales = z.infer<typeof gunnySalesSchema>